implement client_apis for all servers
